в струкруре PB_TempMeasure температура теперь типа float

+++ protobuf/reflow_oven.pb.c
@@ -19,7 +19,7 @@ const pb_field_t PB_Command_fields[4] = {
 
 const pb_field_t PB_TempMeasure_fields[3] = {
     PB_FIELD(  1, UINT32  , SINGULAR, STATIC  , FIRST, PB_TempMeasure, time, time, 0),
-    PB_FIELD(  2, UINT32  , SINGULAR, STATIC  , OTHER, PB_TempMeasure, temp, time, 0),
+    PB_FIELD(  2, FLOAT   , SINGULAR, STATIC  , OTHER, PB_TempMeasure, temp, time, 0),
     PB_LAST_FIELD
 };
